Updated Project (SPA-14068) (SPA-14351) (SPA-13971) (SPA-14425)
Making the update for the Data Extraction project, in order to add the latest version of Sentry, along with the new approach to use it, with the Sentry/Wizard, including the source map and the user information.

Updated the library version, along with the Ag-Grid, MUI versions, date picker update, and products validation.